+++ Code/ObjectMapping/RKObjectMappingProviderContextEntry.m
@@ -43,8 +43,8 @@
 
 - (NSUInteger)hash
 {
-    int prime = 31;
-    int result = 1;
+    NSUInteger prime = 31;
+    NSUInteger result = 1;
     result = prime *[self.userData hash] ? [self.mapping hash] : [self.userData hash];
     return result;
 }
